Fix equal config.xml's that were being reported as different
Would happen if values set had a type different than String. All original
values came from the XML, so they are all strings. Setting something
different would cause it to report that the XML changed.

Not all values are set as string, since it will always become an XML in
the end.
Also, the comparison of the XMLs is made using the stringified version
instead of using the internal parsed hash.

refs #1695