Allow newer versions of BBB using the highest version we support
If a user sets a new version of BBB we don't support yet, now the gem will
use the highest version supported.
This will be useful in most of the cases, since the changes in BBB's API
are usually backwards compatible.
Still raise exceptions if trying to use a version lower than the lowest
supported tough.

refs #1686