Call setConfigXML via POST and change encoding method
Now it uses an encoding method that is more similar to the one used in
BigBlueButton. Fixes issues when there was an asterisk (*) in the
config.xml.

Also made setConfigXML be sent always via POST, so that the XML will be in
the body of the request and not in the URL. It was generating URLs way too
long since the default config.xml is growing.